minor monitoring cleanup
