+++ bothan-core/src/monitoring/types/topic.rs
@@ -1,0 +1,8 @@
+use serde::Serialize;
+
+#[derive(Serialize)]
+#[serde(rename_all = "lowercase")]
+pub enum Topic {
+    Record,
+    Heartbeat,
+}
